WIP. Migrated Jokes module, removed unused modules
